Refactor methods from Contains into Item and Box Services
