+++ src/Main.java
@@ -6,7 +6,7 @@ public class Main {
 
     public static void main(String[] args) {
         ConnectionService connectionService = new ConnectionService();
-        connectionService.connect("ora_j2z8", "a31484116");
+        connectionService.connect("ora_c1z8", "a41584103");
         Session session = new Session();
         CustomerService customerService = new CustomerService(connectionService.getConnection());
         MysteryBoxService mysteryBoxService = new MysteryBoxService(connectionService.getConnection());
